fixing routes
